+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Changelog
 * Add new method :meth:`pycov.CovFrame.matrix_uniformity`.
 * Fix bug in :meth:`pyvcf.slice` method when the input region is missing start or end.
 * Add new command :command:`ngs-bam2fq`.
+* Add new command :command:`fa-filter`.
 
 0.27.0 (2021-11-20)
 -------------------