+++ CHANGELOG.rst
@@ -11,7 +11,7 @@ Changelog
 * Fix bug in :meth:`pyvcf.slice` method when the input region is missing start or end.
 * Add new command :command:`ngs-bam2fq`.
 * Add new command :command:`fa-filter`.
-* Update :meth:`pycov.CovFrame.plot_region` method to raise an error if the CovFrame is empty.
+* Update :meth:`pycov.CovFrame.plot_region` and :meth:`pyvcf.VcfFrame.plot_region` methods to raise an error if the CovFrame/VcfFrame is empty.
 * Update :meth:`pyvcf.VcfFrame.filter_*` methods so that they don't raise an error when the VcfFrame is empty.
 * Update :meth:`common.plot_exons` method to not italicize text by default (use ``name='$text$'`` to italicize).