+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Changelog
 * Update :meth:`pycov.CovFrame.plot_region` and :meth:`pyvcf.VcfFrame.plot_region` methods to raise an error if the CovFrame/VcfFrame is empty.
 * Update :meth:`pyvcf.VcfFrame.filter_*` methods so that they don't raise an error when the VcfFrame is empty.
 * Update :meth:`common.plot_exons` method to not italicize text by default (use ``name='$text$'`` to italicize).
+* Add new argument ``--posix`` to :command:`ngs-hc` command.
 
 0.27.0 (2021-11-20)
 -------------------